pyproject.toml
[project]
name = "freezing-mice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "av>=14.4.0",
    "google-genai>=1.15.0",
    "matplotlib>=3.10.3",
    "moviepy==1.0.3",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "seaborn>=0.13.2",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "google-cloud-storage>=3.1.1",
    "openai>=1.95.0",
    "psutil>=7.0.0",
]
